fix(audit): remove dead RetryPolicy type, add onFailure test, rename spell→graph handler, fix docs
- Remove orphaned RetryPolicy interface from op/types.ts and all re-exports
- Add op-serializer test for onFailure compensation phase rendering
- Rename handlers/spell.ts → handlers/graph.ts; fix stale "spell cast" comment in main.ts
- Fix Temporal lexicon docs: lint-rules count 0→2, strip internal issue ref from worker-profiles.mdx
- Add ops.mdx covering the full Op API and chant run CLI; link from index.mdx
