+++ lexicons/temporal/docs/src/content/docs/worker-profiles.mdx
@@ -3,7 +3,7 @@ title: Worker Profiles
 description: Configure chant.config.ts with Temporal connection profiles for chant run.
 ---
 
-Worker profiles define how `chant run` (issue #8) connects to a Temporal server. They are stored in `chant.config.ts` under the `temporal.profiles` key.
+Worker profiles define how `chant run` connects to a Temporal server. They are stored in `chant.config.ts` under the `temporal.profiles` key.
 
 ## Profile shape
